+++ spec/views/shared/xml/activity_spec.rb
@@ -1,6 +1,4 @@
 RSpec.describe "shared/xml/activity" do
-  let(:activity) { create(:programme_activity) }
-
   before do
     reporting_organisation = build(:partner_organisation)
     render partial: "shared/xml/activity",
@@ -14,17 +12,41 @@ RSpec.describe "shared/xml/activity" do
       }
   end
 
-  context "when there is a commitment" do
-    let!(:commitment) { build(:commitment, activity: activity) }
-    it "renders the commitment" do
-      expect(rendered).to include("<transaction-type code='2'>")
+  describe "commitment" do
+    let(:activity) { create(:programme_activity) }
+
+    context "when there is a commitment" do
+      let!(:commitment) { build(:commitment, activity: activity) }
+      it "renders the commitment" do
+        expect(rendered).to include("<transaction-type code='2'>")
+      end
+    end
+
+    context "when there is no commitment" do
+      let!(:commitment) { nil }
+      it "does not render the commitment" do
+        expect(rendered).not_to include("<transaction-type code='2'>")
+      end
     end
   end
 
-  context "when there is no commitment" do
-    let!(:commitment) { nil }
-    it "does not render the commitment" do
-      expect(rendered).not_to include("<transaction-type code='2'>")
+  describe "activity-scope" do
+    context "when there are benefitting countries" do
+      let(:commitment) { nil }
+      let(:activity) { build(:project_activity, :gcrf_funded, benefitting_countries: %w[ZA]) }
+
+      it "inlcudes the scope element" do
+        expect(rendered).to include("<activity-scope")
+      end
+    end
+
+    context "when there are no benefitting countries" do
+      let(:commitment) { nil }
+      let(:activity) { build(:project_activity, :gcrf_funded, benefitting_countries: nil) }
+
+      it "does not include the scope element" do
+        expect(rendered).not_to include("<activity-scope")
+      end
     end
   end
 end